+++ gtrackProject/Repositories/vehicle/IRepos/IVehicleColorRepository.cs
@@ -7,9 +7,9 @@ namespace gtrackProject.Repositories.vehicle.IRepos
     public interface IVehicleColorRepository
     {
         IQueryable<VehicleColor> GetAll();
-        Task<VehicleColor> Get(int id);
+        Task<VehicleColor> Get(byte id);
         Task<VehicleColor> Add(VehicleColor item);
         Task<bool> Update(VehicleColor item);
-        Task<bool> Remove(int id);
+        Task<bool> Remove(byte id);
     }
 }
